+++ simgui-ds.json
@@ -82,7 +82,7 @@
         47,
         -1,
         -1,
-        -1,
+        39,
         267
       ],
       "povConfig": [
