+++ simgui-window.json
@@ -27,7 +27,7 @@
     "###/SmartDashboard/Auto Mode": {
       "Collapsed": "0",
       "Pos": "162,101",
-      "Size": "363,62"
+      "Size": "361,66"
     },
     "###/SmartDashboard/Elevator Sim": {
       "Collapsed": "0",
@@ -47,12 +47,12 @@
     "###FMS": {
       "Collapsed": "0",
       "Pos": "8,699",
-      "Size": "233,258"
+      "Size": "235,244"
     },
     "###Joysticks": {
       "Collapsed": "0",
       "Pos": "292,408",
-      "Size": "1156,291"
+      "Size": "1156,271"
     },
     "###Keyboard 0 Settings": {
       "Collapsed": "0",
@@ -61,7 +61,7 @@
     },
     "###Keyboard 1 Settings": {
       "Collapsed": "0",
-      "Pos": "1299,27",
+      "Pos": "1299,26",
       "Size": "300,560"
     },
     "###NetworkTables": {
@@ -87,12 +87,12 @@
     "###System Joysticks": {
       "Collapsed": "0",
       "Pos": "0,399",
-      "Size": "273,308"
+      "Size": "273,290"
     },
     "###Timing": {
       "Collapsed": "0",
       "Pos": "1,180",
-      "Size": "199,213"
+      "Size": "197,215"
     },
     "Debug##Default": {
       "Collapsed": "0",
@@ -102,7 +102,7 @@
     "Robot State": {
       "Collapsed": "0",
       "Pos": "5,20",
-      "Size": "143,158"
+      "Size": "137,152"
     }
   }
 }
